requirements.txt
# ==================================================
# Core Python
# ==================================================
python-dotenv>=1.0.1
pyyaml>=6.0
requests>=2.31.0

# ==================================================
# PDF Processing + OCR
# ==================================================
pymupdf>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0

# ==================================================
# Vector Database + Embeddings
# ==================================================
chromadb>=0.4.24
sentence-transformers>=2.6.0
torch>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ==================================================
# Graph Database
# ==================================================
neo4j>=5.18.0
spacy

# ==================================================
# Reranking (Cross Encoder)
# ==================================================
transformers>=4.38.0
tqdm>=4.66.0

# ==================================================
# LLM Integration (Ollama Compatible)
# ==================================================
langchain>=0.1.20
langchain-community>=0.0.38
langchain-ollama>=0.1.0

# ==================================================
# API
# ==================================================
flask>=3.0.0
gunicorn>=21.2.0

# ==================================================
# Authentication & Security
# ==================================================
pyjwt>=2.8.0
bcrypt>=4.1.2

# ==================================================
# Excel Export / Data Handling
# ==================================================
pandas>=2.1.0
openpyxl>=3.1.2

# ==================================================
# Optional Async & Performance
# ==================================================
uvicorn>=0.29.0
httpx>=0.27.0

# ==================================================
# Logging
# ==================================================
loguru>=0.7.2

# ==================================================
# Evaluation / Testing (Optional)
# ==================================================
datasets>=2.18.0
evaluate>=0.4.1